src/main/webapp/app/entities/danh-muc-can-bo/delete/danh-muc-can-bo-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucCanBoService } from '../service/danh-muc-can-bo.service';

import { DanhMucCanBoDeleteDialogComponent } from './danh-muc-can-bo-delete-dialog.component';

describe('DanhMucCanBo Management Delete Component', () => {
  let comp: DanhMucCanBoDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucCanBoDeleteDialogComponent>;
  let service: DanhMucCanBoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucCanBoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucCanBoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucCanBoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucCanBoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
